temp is float not uint8_t tuned PID controller

+++ temp_controller/hardware.h
@@ -3,7 +3,7 @@
 
 void refreshServos();
 void hardwareInit();
-uint8_t getTemp();
+float getTemp();
 uint8_t getGoalTemp();
 void setFaucet(uint8_t value);
 void manualControlMotor();
